### 2024-03-16 #6499 @editorialbot
[PRE REVIEW]: quadkeyr: Tools for converting QuadKeys (Microsoft's Bing Maps Tile System) into raster images
**Submitting author:** @flor14 (<a href="http://orcid.org/0000-0002-0041-097X">María Florencia D&#39;Andrea</a>)
**Repository:** https://github.com/ropensci/quadkeyr
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
