### 2024-03-16 #6499 @osorensen
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.32614/RJ-2018-009 is OK
> - 10.1201/9780429459016 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Bing Maps Tile System - Bing Maps
> - No DOI given, and none found for title: slippymath: Slippy Map Tile Tools
> - Entry without DOI or title found
> - No DOI given, and none found for title: Applied spatial data analysis with R, Second editi...
> - No DOI given, and none found for title: raster: Geographic Data Analysis and Modeling
> - No DOI given, and none found for title: Preparing CRAN for the retirement of rgdal, rgeos ...
> - No DOI given, and none found for title: Edzer Pebesma, Roger Bivand
>
> INVALID DOIs
>
> - None
> ```

@flor14 can you please fix these missing DOIs and notify me here when done?
