### 2024-03-17 dc7a16c
Update paper.md
Use correct bib key
