### 2024-03-17 f6898e9
Update paper.md
https://github.com/openjournals/joss-reviews/issues/6499
