### 2024-03-17 #6500 @editorialbot
[REVIEW]: quadkeyr: Tools for converting QuadKey-identified datasets (Microsoft's Bing Maps Tile System) into raster images and analyzing Meta (Facebook) Mobility Data
**Submitting author:** @flor14 (<a href="http://orcid.org/0000-0002-0041-097X">María Florencia D&#39;Andrea</a>)
**Repository:** https://github.com/ropensci/quadkeyr
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
