### 2024-03-17 fa1ca5e
Update paper.md
The paths to the figures are broken. I think this should fix it, by adding `..` to go up one level. You could test with `@editorialboth generate pdf` in the review issue. If it doesn't work, please move the figures to the `papers` directory.
