+++ paper/paper.md
@@ -38,29 +38,29 @@ the processing of Facebook Mobility data [@data4good]).
 
 The goal of `quadkeyr` is to:
 
-1.  [**Convert a QuadKey to a Simple Features data.frame (and more)**](https://docs.ropensci.org/quadkeyr/articles/quadkey_to_sf_conversion.html) 
+1.  [**Convert a QuadKey to a Simple Features data.frame (and more):**](https://docs.ropensci.org/quadkeyr/articles/quadkey_to_sf_conversion.html) 
 `quadkeyr` provides functions to convert QuadKeys
 to a `sf` POINT data.frame or `sf` POLYGON data.frame [@sf]. 
 Additionally, it offers all the R functions 
 described in the official documentation [@bing-maps]
 for converting QuadKeys to and from tiles, pixels, and geographic coordinates.
 
-![](vignettes/workflow-quadkey.png)
+![](../vignettes/workflow-quadkey.png)
 
-2.  [**Generate Raster Images from Quadkey-Identified    Data**](https://docs.ropensci.org/quadkeyr/articles/quadkey_identified_data_to_raster.html)     Complete a grid of QuadKeys within a specified area and zoom level, 
+2.  [**Generate Raster Images from Quadkey-Identified Data:**](https://docs.ropensci.org/quadkeyr/articles/quadkey_identified_data_to_raster.html) Complete a grid of QuadKeys within a specified area and zoom level, 
 and create a `stars` raster [@stars]. 
 You can also directly convert QuadKeys in a data.frame column 
 into an `sf` POLYGON data.frame.
 
-![](vignettes/workflow-raster.png)
+![](../vignettes/workflow-raster.png)
 
-3. [**Convert Facebook Mobility QuadKey-Identified Datasets into RasterFiles**](https://docs.ropensci.org/quadkeyr/articles/facebook_mobility_csvs_to_raster_files.html)
+3. [**Convert Facebook Mobility QuadKey-Identified Datasets into RasterFiles:**](https://docs.ropensci.org/quadkeyr/articles/facebook_mobility_csvs_to_raster_files.html)
 Convert Facebook mobility data `.csv` files into `.tif` files by day and
 hour reported.
 
-![](vignettes/workflow-facebook.png)
+![](../vignettes/workflow-facebook.png)
 
-4. [**Offer an App for visualizing QuadKeys on a   map**](https://docs.ropensci.org/quadkeyr/articles/quadkey_visualization_app.html)
+4. [**Offer an App for visualizing QuadKeys on a map:**](https://docs.ropensci.org/quadkeyr/articles/quadkey_visualization_app.html)
 Introduce a QuadKey visualization application 
 enabling users to validate function outcomes.
 
@@ -78,7 +78,7 @@ as a sequence of characters \autoref{fig:fig1}.
 They provide an efficient method to address and retrieve specific map tiles, 
 facilitating rapid display within mapping applications.
 
-![The QuadKey of any tile starts with the QuadKey of its parent tile (the containing tile at the previous level). Image extracted from Microsoft’s Bing Maps Tile System webpage. \label{fig:fig1}](vignettes/bing_quadkeys.jpg)
+![The QuadKey of any tile starts with the QuadKey of its parent tile (the containing tile at the previous level). Image extracted from Microsoft’s Bing Maps Tile System webpage. \label{fig:fig1}](../vignettes/bing_quadkeys.jpg)
 
 ## QuadKeys, tiles, pixels and geographic coordinates
 
