### 2024-03-17 #6500 @osorensen
@flor14: the paths to all the figures in the paper seem broken.

For example, the `paper` directory contains the file `bing_tiles.jpg`, but `paper.md` contains the line `![Tile coordinates. Image extracted from Microsoft’s Bing Maps Tile System webpage. \label{fig:fig3}](vignettes/bing_tiles.jpg)`. My last pull request attempts to fix some of this, but I think it's better that you go carefully through all the figures, make corrections and ping me here when done. Feel free to test your code with `@editorialbot generate pdf` in this thread, as much as you like.
