### 2024-03-17 b67220f
Fix papaer.md citations and add JOSS badge
