### 2024-03-17 #6500 @flor14
Thank you, @osorensen. Yes, I am working on improving the citations and fixing the image links.
