### 2024-03-17 54fbe90
delete unecesary images and add size figures
