### 2024-03-17 d96c3c7
Fix images links and sizes inpaper.md
