+++ paper/paper.bib
@@ -1,7 +1,6 @@
 @misc{bing_maps, 
-   title={Bing Maps Tile System - Bing Maps},
-   url={https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}, 
-   journal={Bing Maps | Microsoft Learn Website}, 
+   title={Bing Maps Tile System - Bing Maps. Microsoft Learn Website},
+   howpublished = {\url{https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}}, 
    author={Schwartz, Joe}, 
    year={2022}} 
 @Manual{slippymath,
