+++ paper/paper.bib
@@ -33,7 +33,8 @@
 }
 @misc{data4good, 
 howpublished = {\url{https://dataforgood.facebook.com/}},
-title ={Data For Good at Meta Website}}
+title ={Data For Good at Meta Website},
+year = {2024}}
 @Book{sp,
     author = {Roger S. Bivand and Edzer Pebesma and Virgilio Gomez-Rubio},
     title = {Applied spatial data analysis with {R}, Second edition},
@@ -55,7 +56,7 @@ title ={Data For Good at Meta Website}}
   publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/r-spatial/evolution}},
-  commit = 03612f987016d320687e002992d7e62599fe89d0
+  commit = {03612f987016d320687e002992d7e62599fe89d0}
 }
 @misc{rspatial-evolution-2,
   title={R-spatial evolution: retirement of rgdal, rgeos and maptools},
