### 2024-03-17 2f6a27d
Change image size paper.md
