### 2024-03-17 #6500 @flor14
@osorensen Some of the references are R packages or websites that don't have DOIs assigned. Is it possible to omit this check? How would you suggest proceeding?
