### 2024-03-17 42d44b3
Update paper.md
I think it makes sense with colons after the headers in the bullet list.
