### 2024-03-17 ropensci/quadkeyr#14 @osorensen
Editorial comment on JOSS submission
I cannot see that Figure 4 is referenced anywhere in the text. Could you please add this?

https://github.com/openjournals/joss-reviews/issues/6500
