### 2024-03-17 17ff5b8
Update paper.md
I think it makes more sense to say "the" than "both" here.

https://github.com/openjournals/joss-reviews/issues/6500
