### 2024-03-17 ropensci/quadkeyr#16 @osorensen
Editorial comment on JOSS submission
The name "Roger Bivand" seems to be spelled differently across your references in you BibTeX source file. This makes them appear as "R. Bivand" and "R. S. Bivand", as shown in the screenshot below. Lease go through `paper.bib` and make sure there is only one spelling, since "R. Bivand" and "R. S. Bivand" are the same person.

https://github.com/openjournals/joss-reviews/issues/6500
