### 2024-03-17 662542d
Update paper.bib
I think at least Bing should appear with uppercase "B" in the reference list, so suggest protecting this with curly braces.

https://github.com/openjournals/joss-reviews/issues/6500
