### 2024-03-17 #6500 @osorensen
@flor14, as you see I have opened a number of pull requests and issue based on a new reading of the paper. All very minor edits. Please let me know in this thread when they're fixed.
