### 2024-03-18 e2b6df1
Fix autor paper.bib
