+++ paper/paper.bib
@@ -36,7 +36,7 @@ howpublished = {\url{https://dataforgood.facebook.com/}},
 title ={Data For Good at Meta Website},
 year = {2024}}
 @Book{sp,
-    author = {Roger S. Bivand and Edzer Pebesma and Virgilio Gomez-Rubio},
+    author = {Roger Bivand and Edzer Pebesma and Virgilio Gomez-Rubio},
     title = {Applied spatial data analysis with {R}, Second edition},
     year = {2013},
     publisher = {Springer, NY},
