### 2024-03-18 5414c1d
Add changes in citations paper and update CITATION and codemeta
