+++ paper/paper.bib
@@ -62,7 +62,7 @@ year = {2024}}
   title={R-spatial evolution: retirement of rgdal, rgeos and maptools},
   url={https://r-spatial.org/r/2022/04/12/evolution.html#packages-depending-on-sp-and-raster}, 
   journal={r-spatial.org}, 
-  author={Pebesma, Edzer and Bivand, Roger}, 
+  author={Edzer Pebesma and Roger Bivand}, 
   year={2022}, 
   month={Apr}} 
 @Manual{rgdal,
