### 2024-03-18 ropensci/quadkeyr#16 @osorensen
This seems to be fixed in the latest version.
