### 2024-03-18 #6500 @osorensen
@flor14, could you please also fix [this issue](https://github.com/ropensci/quadkeyr/issues/14)?
