### 2024-03-18 f1f1303
Minor changes paper.md
