### 2024-03-18 b3a6079
Fix figure reference paper.md
