### 2024-03-18 3bfd62d
Update paper.md
Figure captions should be complete sentences.

https://github.com/openjournals/joss-reviews/issues/6500
