### 2024-03-18 f8ece5f
Update paper.md
More minor edits.

https://github.com/openjournals/joss-reviews/issues/6500
