+++ paper/paper.md
@@ -120,8 +120,10 @@ the structure of Tile Maps. All of these intermediary functions,
 as well as those facilitating the reverse conversion, 
 are available for use within the `quadkeyr` package.
 
-There are different tools in Python, R and other programming languages 
-that help to convert Quadkeys to tiles, pixel or geographic coordinates as well. 
+There are different tools in Python, 
+R and other programming languages 
+that help to convert Quadkeys to 
+tiles, pixel or geographic coordinates as well. 
 The package `slippymath` [@slippymath] is the sole package
 in the R ecosystem dedicated to analyzing tile maps. 
 Despite some similarities 
@@ -129,11 +131,12 @@ between `quadkeyr` and `slippymath`,
 the packages were developed with different objectives.
 
 The goal of `quadkeyr` is 
-to provide a new tool to facilitate the analysis of Meta (Facebook) Mobility data, 
+to provide a new tool to facilitate 
+the analysis of Meta (Facebook) Mobility Data, 
 so it has specific functions tailored for this purpose
-and is based solely on the Microsoft's Bing Tile Maps documentation.
+and is based solely on Microsoft's Bing Tile Maps documentation.
 On the other hand, `slippymath` is a mature and smaller package 
-with general tools for analyzing tile maps.
+featuring general functions for processing tile maps.
 It was designed to be easily incorporated as a 
 dependency of higher-level tools.
 
@@ -152,7 +155,7 @@ which converts the `sf` POLYGON data.frame provided as input
 to a `stars` object.
 This object can then be saved as a raster image.
 
-## Meta (Facebook) mobility data and Data for Good
+## Meta (Facebook) Mobility Data and Data for Good
 
 Meta (formerly Facebook) provides datasets
 in their program Data for good 
@@ -179,6 +182,8 @@ We want to acknowledge the ROpenSci community,
 especially Emily Riederer, Maria Paula Caldas,
 Vincent van Hees, and Miles McBain, 
 for their valuable comments and support
-during the review process.
+during the review process, 
+and also Øystein Sørensen
+for proofreading of this manuscript.
 
 # References
