### 2024-03-18 bbbb501
Update paper.bib
Reference should be to book, not to single page.

https://github.com/openjournals/joss-reviews/issues/6500
