+++ paper/paper.bib
@@ -27,8 +27,7 @@
   year = {2023},
   publisher = {Chapman and Hall/CRC},
   address = {London},
-  url = {https://r-spatial.org/book/},
-  pages = {352},
+  url = {https://r-spatial.org/book/},  
   doi = {10.1201/9780429459016},
 }
 @misc{data4good, 
