### 2024-03-18 #6500 @flor14
@osorensen Thank you for all the updates.
My latest changes have been:
- Minor modifications in the `slippymath` paragraph (lines 53 - 63)
- Homogenization of the different ways of writing Facebook Mobility Data regarding capitalization in all the manuscript.
- Modifying my affiliation and adding Pilar as a corresponding author.
- Minor changes Acknowledgements section and Keywords.

I've read everything again and it seems OK.
