+++ CITATION.cff
@@ -8,7 +8,7 @@ message: 'To cite package "quadkeyr" in publications use:'
 type: software
 license: MIT
 title: 'quadkeyr: Tools for converting QuadKey-identified datasets (Microsoft''s Bing
-  Maps Tile System) into raster images and analyzing Meta (Facebook) mobility data.'
+  Maps Tile System) into raster images and analyzing Meta (Facebook) Mobility Data.'
 version: 0.1.0
 abstract: '''Quadkeyr'' functions generate raster images based on QuadKey-identified
   data, facilitating efficient integration of Tile Maps data into R workflows. In
