+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: quadkeyr
-Title: Tools for converting QuadKey-identified datasets (Microsoft's Bing Maps Tile System) into raster images and analyzing Meta (Facebook) mobility data.
+Title: Tools for converting QuadKey-identified datasets (Microsoft's Bing Maps Tile System) into raster images and analyzing Meta (Facebook) Mobility Data.
 Version: 0.1.0
 Authors@R: 
     c(person(given = "Florencia",
@@ -27,7 +27,7 @@ Authors@R:
     person("CDC's Center for Forecasting and Outbreak Analytics",
             comment = c("This project was made possible by cooperative agreement CDC-RFA-FT-23-0069 (grant # NU38FT000009-01-00) from the CDC's Center for Forecasting and Outbreak Analytics. Its contents are solely the responsibility of the authors and do not necessarily represent the official views of the Centers for Disease Control and Prevention."),  role = c("fnd")),
     person("MIDAS-NIH COVID-19 urgent grant program", role = c("fnd")),
-    person("Washington State University", role = c("cph")))
+    person("Paul G. Allen School for Global Health, Washington State University", role = c("cph")))
 Description: 'Quadkeyr' functions generate raster images based on QuadKey-identified data, facilitating efficient integration of Tile Maps data into R workflows. In particular, 'Quadkeyr' provides support to process and analyze Facebook mobility datasets within the R environment.
 License: MIT + file LICENSE
 URL: https://docs.ropensci.org/quadkeyr/, https://github.com/ropensci/quadkeyr
