### 2024-03-18 #6500 @osorensen
> Not yet, I was considering doing it after the paper was ready.

@flor14, I missed this one, but I think the paper is ready now. In order to proceed with acceptance, the next step is that you create a Zenodo archive.
