### 2024-03-19 e1bc699
Make menu titles shorter in the docs
