### 2024-03-19 #6500 @flor14
These missing DOIs are from software packages or websites that don't have DOIs assigned.
