### 2024-03-20 cc3f66f
Try to fix long size titles in docs menu
