+++ README.Rmd
@@ -59,7 +59,7 @@ The goal of `quadkeyr` is to:
 `quadkeyr` provides functions to convert a QuadKey to
 a `sf` POINT data.frame or `sf` POLYGON data.frame.
 Additionally, it offers all the R functions 
-described in the official documentation in the [official documentation](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system) for converting QuadKeys to and from tiles, pixels, and geographic coordinates.
+described in the [official documentation](https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system) for converting QuadKeys to and from tiles, pixels, and geographic coordinates.
 
 ```{r w1, echo = FALSE, out.width= "80%", fig.align='center'}
 knitr::include_graphics("vignettes/workflow_quadkey.png")
