### 2024-03-21 05973bc
Add shorter titles to pkgdown menu
