### 2024-03-25 #6500 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
